Assignemnt1/requirements.txt
torch
torchvision
fastapi
uvicorn
python-multipart
Pillow
